% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCleanLmData.R
\name{GetCleanLmData}
\alias{GetCleanLmData}
\title{Get subset data containing certain attributes, which can be used for analyses such as
linear regressions.}
\usage{
GetCleanLmData(data, y, x, FE, common.na.strings, ...)
}
\arguments{
\item{data}{data containing variables to be analyzed, default is GSS.}

\item{y}{variable containing outcome variable in data, has to be numeric.}

\item{x}{variable containing treatment status in data, has to be indicator having two levels.}

\item{FE}{fixed effects to be included.}

\item{common.na.strings}{values where to be replaced by NA}

\item{...}{control variables to be selected from data.}
}
\value{
Clean data with selected outcome, treatment, fixed effects, and control variables.
        The data can be used for analyses like linear regressions.
}
\description{
This function selects subset of data including outcome variable of interest, treatment of interest,
fixed effects. Users can also select other control variables affecting the selected outcome variable.
}
\examples{
data(GSS)
GetCleanLmData(GSS,'happy.score','election','year',
c("DK", "IAP", "UNCODEABLE & IAP", ".a", ".b","DK NA"),'age','income')
}
