% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSS.R
\docType{data}
\name{GSS}
\alias{GSS}
\title{Data on household happiness, income and other characteristics from `gss.norc.org`.}
\description{
This data is processed based on raw data downloaded from `gss.norc.org``, which can be found
in folder 'data-raw'. Details of data preprocessing is included in `/data-raw/data_preprocessing.R`.
The dta consists of a named list of variables including `happy` and `year`, which are the
variables of interests. Other attributes such as `age`, `income` are related control variables.
}
\references{
[gss.norc.org
    Data](http://gss.norc.org/get-the-data/stata)
}
\keyword{data}
