% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagEffect.R
\name{LagEffect}
\alias{LagEffect}
\title{Event study: lag effect.}
\usage{
LagEffect(data, y, x, time, FE)
}
\arguments{
\item{data}{data containing variables to be analyzed.}

\item{y}{outcome variable.}

\item{x}{treatment variable.}

\item{time}{time fixed effect.}

\item{FE}{fixed effects to to included.}
}
\value{
summary of lm fit on lag treatment.
}
\description{
This functions examines if there is a lag effect on the relationship between x and y.
}
\examples{
data(GSS)
LmData <- GetCleanLmData (GSS,'happy.score', 'election', 'year', 'age', 'income')
LagEffect(LmData,'happy.score','election','year', 'year')
}
