% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeadEffect.R
\name{LeadEffect}
\alias{LeadEffect}
\title{Event study: lead effect.}
\usage{
LeadEffect(data, y, x, time, FE)
}
\arguments{
\item{data}{data containing variables to be analyzed.}

\item{y}{outcome variable.}

\item{x}{treatment variable.}

\item{time}{time fixed effect.}

\item{FE}{fixed effects to to included.}
}
\value{
summary of regression on lead treatment.
}
\description{
This function examines if there is lead effect on the relationship between x and y.
}
\examples{
data(GSS)
LmData <- GetCleanLmData (GSS,'happy.score', 'election', 'year', 'age', 'income')
LeadEffect(LmData,'happy.score','election','year', 'year')
}
