% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlaceboTSReshuffle.R
\name{PlaceboTSReshuffle}
\alias{PlaceboTSReshuffle}
\title{Placebo test on effect of x on y.}
\usage{
PlaceboTSReshuffle(data, y, x, FE)
}
\arguments{
\item{data}{data containing variables to be analyzed, default is GSS.}

\item{y}{variable containing outcome variable in data.}

\item{x}{variable containing treatment status in data, has to be indicator having two levels.}

\item{FE}{fixed effects.}
}
\value{
A plot of coefficients from re-shuffled and original treatment status.
}
\description{
This function executes a pacebo test on effect of x on y.
Specifically, treatment status is randomly re-shuffled for 100 times, and the effect is re-examined on each
re-shuffling. The coefficients from re-shuffled and original treatment status are ploted.
If the effect is robust, then the point from original treatment status should be an outlier.
}
\examples{
data(GSS)
LmData <- GetCleanLmData (GSS,'happy.score', 'election', 'year', 'age', 'income')
PlaceboTSReshuffle(LmData,'happy.score','election','year')
}
